Modify: db user infos move to .env
-adjust file names
-db info use .env
