+++ model/database_conn.py
@@ -1,9 +1,13 @@
+import os
+from dotenv import load_dotenv
 import mysql.connector
 from mysql.connector import pooling
 
+load_dotenv()
+
 MysqlConnectInfo = {
-	"user" : "root",
-	"password" : "root123",
+	"user" : os.getenv("DB_USERNAME"),
+	"password" : os.getenv("DB_PASSWORD"),
 	"host" : "localhost",
 	"database" : "TaipeiAttr"
 }
